Feature

+++ Android/App/Mix_Manual/src/main/java/com/zktony/www/ui/home/ProgramExecutor.kt
@@ -10,8 +10,8 @@ import kotlinx.coroutines.*
  * @param scope 协程作用域
  */
 class ProgramExecutor constructor(
-    private val colloid: Int,
-    private val coagulant: Int,
+    private val colloid: Float,
+    private val coagulant: Float,
     private val mode: Boolean,
     private val scope: CoroutineScope = CoroutineScope(Dispatchers.IO)
 ) {
@@ -25,7 +25,7 @@ class ProgramExecutor constructor(
                     step {
                         v1 = if (mode) (colloid / 2f - 1f) else (colloid / 2f + 1f)
                         v2 = if (mode) (colloid / 2f + 1f) else (colloid / 2f - 1f)
-                        v3 = coagulant.toFloat()
+                        v3 = coagulant
                     }
                 }
             }
@@ -44,9 +44,9 @@ class ProgramExecutor constructor(
                 execute {
                     type(0)
                     step {
-                        v1 = if (mode) 0f else colloid.toFloat()
-                        v2 = if (mode) colloid.toFloat() else 0f
-                        v3 = coagulant.toFloat()
+                        v1 = if (mode) 0f else colloid
+                        v2 = if (mode) colloid else 0f
+                        v3 = coagulant
                     }
                 }
             }
